pyproject.toml
[tool.poetry]
name = "kindgen"
version = "0.1.0"
description = ""
authors = ["Jonah Ramponi <jonah.ramponi@smartr.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cohere = "^5.9.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
